feat : report_api connection

+++ Android/app/src/main/java/com/example/myapplication/data/dto/response/care/ReportDto.kt
@@ -1,6 +1,6 @@
 package com.example.myapplication.data.dto.response.care
 
 data class ReportDto(
-    val reportId: Int,
+    val reportId: Long,
     val createdAt: Long
 )
